Added console.log to ownerByAlbum prtected url.

+++ app/controllers/albumController.js
@@ -142,7 +142,7 @@ const editAlbum = async (req, res, next) => {
 
         const {tracks} = req.body;
 
-        const albumToEdit = await Album.find({"id": albumId});
+        const albumToEdit = await Album.findOne({"id": albumId});
 
         if (albumToEdit === null) {
             next(ApiError.resourceNotFound(`Album with ${albumId} not found`))
@@ -150,21 +150,28 @@ const editAlbum = async (req, res, next) => {
         }
 
         // Update to the albums tracks
-        const albumSongsId = albumToEdit.tracks.map((track) => track._id);
-        const songsToDelete = albumSongsId.filter(x => !tracks.includes(x));
-        albumToEdit.tracks.apply((song) => {
-            if (songsToDelete.includes(song._id)) {
-                song.remove();
-            }
-        });
+        let songsToDelete = [];
+        if (tracks) {
+            const albumSongsId = albumToEdit.tracks.map((track) => track._id);
+            songsToDelete = albumSongsId.filter(x => !tracks.includes(x));
+            albumToEdit.tracks.apply((song) => {
+                if (songsToDelete.includes(song._id)) {
+                    song.remove();
+                }
+            });
+
+            delete req.body.tracks;
+        }
+
 
         // Update to the album values
-        delete req.body.tracks;
         Object.assign(albumToEdit, req.body);
         await albumToEdit.save();
 
         // Delete dependencies of deleted songs
-        await deleteForeignKeys(songsToDelete);
+        if (songsToDelete) {
+            await deleteForeignKeys(songsToDelete);
+        }
 
         res.status(204).send({})
     } catch (err) {
@@ -178,7 +185,7 @@ const deleteAlbum = async (req, res, next) => {
     const albumId = req.params.id
     try {
 
-        const albumToDelete = await Album.find({"id": albumId});
+        const albumToDelete = await Album.findOne({"id": albumId});
         if (albumToDelete === null) {
             next(ApiError.resourceNotFound(`Album with ${albumId} not found`))
             return;
